Fixed conflicts:
	src/Array3D.cpp

+++ src/Array3D.cpp
@@ -134,22 +134,22 @@ Array3D<T>& Array3D<T>::operator=(const Array3D<T> &&a)
 
 // DATA ACCESS AND MODIFICATION FUNCTIONS
 template <class T>
-int Array3D<T>::GetDim(int dim) const
+size_t Array3D<T>::GetDim(int dim) const
 {
-#ifndef RELEASE
-    assert(dim > 0 && dim < 4);
-    size_t retval = 0;
-    if(dim == 1){retval = size1;}
-    if(dim == 2){retval = size2;}
-    if(dim == 3){retval = size3;}
-    return retval;
-#else
-    size_t retval = 0;
-    if(dim == 1){retval = size1;}
-    if(dim == 2){retval = size2;}
-    if(dim == 3){retval = size3;}
-    return retval;
-#endif
+	#ifndef RELEASE
+		assert(dim > 0 && dim < 4);
+        size_t retval = 0;
+		if(dim == 1){retval = size1;}
+		if(dim == 2){retval = size2;}
+		if(dim == 3){retval = size3;}
+		return retval;
+	#else
+        size_t retval = 0;
+		if(dim == 1){retval = size1;}
+		if(dim == 2){retval = size2;}
+		if(dim == 3){retval = size3;}
+		return retval;
+	#endif
 }
 
 template <class T> inline
